updates to validation, main menue added to navigate more options

+++ Game.cs
@@ -9,6 +9,7 @@ namespace LemonadeStandV2
         List<Day> days;
         int currentDay;
         int gameLength;
+        bool quitGame;
 
         public Game()
         {
@@ -34,16 +35,29 @@ namespace LemonadeStandV2
                 {
                     Console.Clear();
 
-                    UserInterface.DisplayDay(currentDay);
-                    UserInterface.DisplayForcast(days[i]);
-                    UserInterface.SevenDayForcast(days, days[i], currentDay);
-                    UserInterface.DisplayInventory(player);
-                    GoToStore();
-                    player.recipe.GoToRecipe();
-                    RunDay();  
-                    player.inventory.iceCubes.RemoveRange(0, player.inventory.iceCubes.Count);
-                    UserInterface.DisplayPostInventory(player);
-                    currentDay++;
+                    UserInterface.DisplayDay(currentDay, days);
+                    //UserInterface.DisplayForcast(days[i]);
+                    //UserInterface.SevenDayForcast(days, days[i], currentDay);
+                    //UserInterface.DisplayInventory(player);
+
+                    bool start;
+                    do
+                    {
+                        GoToMenu();
+                        start = GoodToStart();
+                    }
+                    while (start == false);
+
+                    if (quitGame == false)
+                    {
+
+                        RunDay();
+                        player.inventory.iceCubes.RemoveRange(0, player.inventory.iceCubes.Count);
+                        UserInterface.DisplayPostInventory(player);
+                        currentDay++;
+                    }
+
+                   
                 }
                 EndGame();
                 playAgain = UserInterface.PlayAgainMenu();
@@ -68,7 +82,74 @@ namespace LemonadeStandV2
             days.Add(day);        }        public void CreateAllDays()        {            days = new List<Day> { };            for (int i = 1; i <= gameLength; i++)            {                NewDay(i);
             }        }
 
+        public void GoToMenu()
+        {
 
+            
+            UserInterface.DisplayForcast(days[currentDay-1]);
+            UserInterface.DisplayInventory(player);
+            UserInterface.DisplayRecipe(player);
+
+            bool leaveMenu = false;
+            while (leaveMenu == false)
+            {
+                UserInterface.DisplayMenu();
+                string selection = UserInterface.GetUserInput("Where would you like to go?");
+                while (selection.ToLower() != "start" && selection.ToLower() != "store" && selection.ToLower() != "recipe" && selection.ToLower() != "quit" && selection.ToLower() != "forcast")
+                {
+                    selection = UserInterface.RetryGetUserInput("not a valid selection!");
+                }
+                switch (selection)
+                {
+                    case "store":
+                        GoToStore();
+                        break;
+                    case "recipe":
+                        player.recipe.GoToRecipe();
+                        break;
+                    case "start":
+                        leaveMenu = true;
+                        break;
+                    case "quit":
+                        leaveMenu = true;
+                        quitGame = true;
+                        break;
+                    case "forcast":
+                        UserInterface.SevenDayForcast(days, currentDay - 1);
+                        break;
+                }
+            }
+            }
+
+        public bool GoodToStart()
+        {
+            if (player.inventory.cups.Count == 0)
+            {
+                UserInterface.Error("WOAH! you have no cups in your inventory! You cant sell ANY lemonade without 12 cups! Check your inventory and go back to the store!");
+                return false;
+            }
+            if (player.recipe.amountOfLemons == 0 || player.recipe.amountOfSugarCups == 0 || player.recipe.amountOfIceCubes == 0)
+            {
+                
+                string sure = UserInterface.YesOrNo(UserInterface.RetryGetUserInput("One or more items in your recipe have a quantity of ZERO, proceed 'yes' or 'no'?"));
+                if (sure == "yes")
+                {
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
+            }
+            if (player.inventory.lemons.Count < player.recipe.amountOfLemons || player.inventory.sugarCups.Count < player.recipe.amountOfSugarCups || player.inventory.iceCubes.Count < player.recipe.amountOfIceCubes)
+            {
+                UserInterface.Error("You have insufficient inventory to use your current recipee! change the recipe or go back to the store!");
+                return false;
+            }
+
+
+            return true;
+        }
 
 
         public void GoToStore()
@@ -120,7 +201,7 @@ namespace LemonadeStandV2
 
             player.wallet.Money += dailyGross;
             UserInterface.DisplayDayResult(cupsSold, dailyGross, days[currentDay-1], currentDay);
-            Thread.Sleep(2000);
+            
 
         }
